chore: move inbounds/outbounds related things to connections module
