+++ src/app/connections/README.md
@@ -1,0 +1,4 @@
+# connections
+
+A 'connection' refers to an `inbound` or `outbound` of a dataplane proxy.
+
