+++ src/app/connections/components/connection-traffic/ConnectionCard.vue
