+++ src/app/connections/components/connection-traffic/ConnectionGroup.vue
