+++ src/app/connections/components/connection-traffic/README.md
@@ -1,0 +1,4 @@
+# ConnectionTraffic
+
+Component for displaying traffic across `inbounds` or `outbounds`.
+
