+++ src/app/connections/index.ts
@@ -1,6 +1,4 @@
-import { features } from './features'
 import { sources } from './sources'
-import { services as connections } from '@/app/connections'
 import type { ServiceDefinition } from '@/services/utils'
 import { token } from '@/services/utils'
 export * from './routes'
@@ -9,26 +7,15 @@ type Token = ReturnType<typeof token>
 
 export const services = (app: Record<string, Token>): ServiceDefinition[] => {
   return [
-    [token('data-planes.sources'), {
+    [token('connections.sources'), {
       service: sources,
       arguments: [
         app.source,
         app.api,
-        app.can,
       ],
       labels: [
         app.sources,
       ],
     }],
-    [token('data-planes.features'), {
-      service: features,
-      arguments: [
-        app.env,
-      ],
-      labels: [
-        app.features,
-      ],
-    }],
-    ...connections(app),
   ]
 }
