+++ src/app/connections/views/ConnectionOutboundSummaryClustersView.vue
@@ -9,20 +9,20 @@
       dataPlane: '',
       service: '',
     }"
-    name="data-plane-outbound-summary-stats-view"
+    name="connection-outbound-summary-clusters-view"
   >
     <AppView>
       <template #title>
         <h3>
           <RouteTitle
-            :title="`Stats`"
+            :title="`Clusters`"
           />
         </h3>
       </template>
       <div>
         <DataSource
-          v-slot="{ data, error, refresh }: StatsSource"
-          :src="`/meshes/${route.params.mesh}/dataplanes/${route.params.dataPlane}/stats/${route.params.service}`"
+          v-slot="{ data, error, refresh }: ClustersDataSource"
+          :src="`/meshes/${route.params.mesh}/dataplanes/${route.params.dataPlane}/data-path/clusters`"
         >
           <ErrorBlock
             v-if="error"
@@ -34,11 +34,12 @@
           <CodeBlock
             v-else
             language="json"
-            :code="(() => data.raw.split('\n')
-              .filter((item) => item.includes(`.${route.params.service}.`))
-              .map((item) => item.replace(`${route.params.service}.`, ''))
-              .join('\n')
-            )()"
+            :code="(() => `${
+              data.split('\n')
+                .filter(item => item.startsWith(`${route.params.service}::`))
+                .map(item => item.replace(`${route.params.service}::`, ''))
+                .join('\n')
+            }`)()"
             is-searchable
             :query="route.params.codeSearch"
             :is-filter-mode="route.params.codeFilter"
@@ -66,8 +67,8 @@
 import { KUI_ICON_SIZE_30 } from '@kong/design-tokens'
 import { RefreshIcon } from '@kong/icons'
 
-import { StatsSource } from '../sources'
 import CodeBlock from '@/app/common/code-block/CodeBlock.vue'
 import ErrorBlock from '@/app/common/ErrorBlock.vue'
 import LoadingBlock from '@/app/common/LoadingBlock.vue'
+import type { ClustersDataSource } from '@/app/data-planes/sources'
 </script>
