+++ src/app/data-planes/routes.ts
@@ -1,6 +1,8 @@
+import { routes as connections } from '@/app/connections/routes'
 import type { RouteRecordRaw } from 'vue-router'
 export const routes = () => {
   const item = (): RouteRecordRaw[] => {
+    const children = connections()
     return [
       {
         path: 'data-planes/:dataPlane',
@@ -11,53 +13,7 @@ export const routes = () => {
             path: 'overview',
             name: 'data-plane-detail-view',
             component: () => import('@/app/data-planes/views/DataPlaneDetailView.vue'),
-            children: [
-              {
-                path: 'inbound/:service',
-                name: 'data-plane-inbound-summary-view',
-                component: () => import('@/app/data-planes/views/DataPlaneInboundSummaryView.vue'),
-                children: [
-                  {
-                    path: 'overview',
-                    name: 'data-plane-inbound-summary-overview-view',
-                    component: () => import('@/app/data-planes/views/DataPlaneInboundSummaryOverviewView.vue'),
-                  },
-                  {
-                    path: 'stats',
-                    name: 'data-plane-inbound-summary-stats-view',
-                    component: () => import('@/app/data-planes/views/DataPlaneInboundSummaryStatsView.vue'),
-                  },
-                  {
-                    path: 'clusters',
-                    name: 'data-plane-inbound-summary-clusters-view',
-                    component: () => import('@/app/data-planes/views/DataPlaneInboundSummaryClustersView.vue'),
-                  },
-                ],
-              },
-              {
-                path: 'outbound/:service',
-                name: 'data-plane-outbound-summary-view',
-                component: () => import('@/app/data-planes/views/DataPlaneOutboundSummaryView.vue'),
-                children: [
-                  {
-                    path: 'overview',
-                    name: 'data-plane-outbound-summary-overview-view',
-                    component: () => import('@/app/data-planes/views/DataPlaneOutboundSummaryOverviewView.vue'),
-                  },
-                  {
-                    path: 'stats',
-                    name: 'data-plane-outbound-summary-stats-view',
-                    component: () => import('@/app/data-planes/views/DataPlaneOutboundSummaryStatsView.vue'),
-                  },
-                  {
-                    path: 'clusters',
-                    name: 'data-plane-outbound-summary-clusters-view',
-                    component: () => import('@/app/data-planes/views/DataPlaneOutboundSummaryClustersView.vue'),
-                  },
-
-                ],
-              },
-            ],
+            children,
           },
           {
             path: 'policies',
